+++ e2e/integration/frontity-01/google-tag-manager-analytics.spec.js
@@ -27,28 +27,28 @@ describe("Google Tag Manager", () => {
   });
 
   it("should have sent the first pageview", () => {
-    cy.window().its("dataLayer").its(0).should("deep.equal", pageviewHome);
+    cy.window().its("dataLayer").its(1).should("deep.equal", pageviewHome);
   });
 
   it("should sent a pageview if the page changes", () => {
     cy.get("button#change-link").click();
-    cy.window().its("dataLayer").its(1).should("deep.equal", pageviewSomePost);
+    cy.window().its("dataLayer").its(2).should("deep.equal", pageviewSomePost);
   });
 
   it("should sent pageviews when going back or forward", () => {
     cy.get("button#change-link").click();
     cy.go("back");
-    cy.window().its("dataLayer").its(2).should("deep.equal", pageviewHome);
+    cy.window().its("dataLayer").its(3).should("deep.equal", pageviewHome);
 
     cy.go("forward");
-    cy.window().its("dataLayer").its(3).should("deep.equal", pageviewSomePost);
+    cy.window().its("dataLayer").its(4).should("deep.equal", pageviewSomePost);
   });
 
   it("should send events", () => {
     // Wait for the first pageview to be sent.
-    cy.window().its("dataLayer").its(0).should("deep.equal", pageviewHome);
+    cy.window().its("dataLayer").its(1).should("deep.equal", pageviewHome);
     // Send event.
     cy.get("button#send-event").click();
-    cy.window().its("dataLayer").its(1).should("deep.equal", someEvent);
+    cy.window().its("dataLayer").its(2).should("deep.equal", someEvent);
   });
 });