+++ e2e/project/frontity.settings.ts
@@ -129,6 +129,18 @@ const settings: Settings = [
     ],
   },
   {
+    name: "wp-comments",
+    packages: [
+      "e2e-wp-comments",
+      "@frontity/tiny-router",
+      "@frontity/wp-comments",
+      {
+        name: "@frontity/wp-source",
+        state: { source: { api: "http://localhost:8080/wp-json" } },
+      },
+    ],
+  },
+  {
     name: "google-ad-manager",
     packages: [
       "e2e-ads",
@@ -187,6 +199,95 @@ const settings: Settings = [
         state: {
           source: {
             api: "http://localhost:8080/wp-json",
+          },
+        },
+      },
+    ],
+  },
+  {
+    name: "yoast-package",
+    state: {
+      frontity: {
+        url: "http://my.frontity.site",
+      },
+    },
+    packages: [
+      "@frontity/tiny-router",
+      "@frontity/html2react",
+      {
+        name: "@frontity/wp-source",
+        state: {
+          source: {
+            api: "http://localhost:8080/wp-json",
+            postTypes: [
+              {
+                type: "movie",
+                endpoint: "movies",
+                archive: "/movies",
+              },
+            ],
+            taxonomies: [
+              {
+                taxonomy: "actor",
+                endpoint: "actors",
+                postTypeEndpoint: "movies",
+              },
+            ],
+          },
+        },
+      },
+      {
+        name: "@frontity/yoast",
+        state: {
+          yoast: {
+            renderTags: "server",
+          },
+        },
+      },
+    ],
+  },
+  {
+    name: "head-tags",
+    state: {
+      frontity: {
+        url: "http://my.frontity.site",
+      },
+    },
+    packages: [
+      "@frontity/tiny-router",
+      "@frontity/html2react",
+      {
+        name: "@frontity/wp-source",
+        state: {
+          source: {
+            api: "http://localhost:8080/wp-json",
+            postTypes: [
+              {
+                type: "movie",
+                endpoint: "movies",
+                archive: "/movies",
+              },
+            ],
+            taxonomies: [
+              {
+                taxonomy: "actor",
+                endpoint: "actors",
+                postTypeEndpoint: "movies",
+              },
+            ],
+            params: {
+              ["head_tags_skip_cache"]: "true",
+            },
+          },
+        },
+      },
+      {
+        name: "@frontity/head-tags",
+        state: {
+          headTags: {
+            transformLinks: {
+              base: "http://localhost:8080",
+            },
           },
         },
       },