+++ e2e/support/commands.js
@@ -1,6 +1,6 @@
 /* eslint-disable jest/valid-expect */
 
-import "cypress-plugin-snapshots/commands";
+require("cypress-plugin-snapshots/commands");
 
 Cypress.Commands.add("isInViewport", { prevSubject: true }, (subject) => {
   const bottom = Cypress.$(cy.state("window")).height();
@@ -20,4 +20,31 @@ Cypress.Commands.add("isNotInViewport", { prevSubject: true }, (subject) => {
   expect(rect.bottom).to.be.greaterThan(bottom);
 
   return subject;
+});
+
+/**
+ * Extends the `visit` command to add an option to disable scripts. Useful when
+ * testing SSR content.
+ *
+ * Based on the solution @pke proposed on the following comment:
+ * https://github.com/cypress-io/cypress/issues/1611#issuecomment-623896822.
+ */
+Cypress.Commands.overwrite("visit", (visit, url, options = {}) => {
+  const parentDocument = cy.state("window").parent.document;
+  const iframe = parentDocument.querySelector(".iframes-container iframe");
+  if (false === options.script) {
+    if (false !== Cypress.config("chromeWebSecurity")) {
+      throw new TypeError(
+        "When you disable script you also have to set 'chromeWebSecurity' in your config to 'false'"
+      );
+    }
+    iframe.sandbox = "";
+  } else {
+    /**
+     * In case it was added by a visit before, the attribute has to be removed
+     * from the iframe.
+     */
+    iframe.removeAttribute("sandbox");
+  }
+  return visit(url, options);
 });