+++ packages/wp-source/src/state.ts
@@ -10,6 +10,16 @@ const state: WpSource["state"]["source"] = {
     }
     const { route, query, page } = parse(link);
 
+    // The non-URL resources start with `@` by convention. In that case,
+    // it makes no sense to add the `link`, `route` and `query`
+    if (link.startsWith("@")) {
+      return {
+        page,
+        isFetching: false,
+        isReady: false,
+      };
+    }
+
     return {
       link: normalize(normalizedLink),
       route,
@@ -18,6 +28,32 @@ const state: WpSource["state"]["source"] = {
       isFetching: false,
       isReady: false,
     };
+  },
+  entity: ({ state }) => (link) => {
+    // Get the data object pointed by `link`.
+    const data = state.source.get(link);
+
+    // Initialize entity as `null` (it is possible that data doesn't point to an
+    // entity, e.g. a date archive or a 404 page).
+    let entity: any = null;
+
+    // Entities are stored in different places depending on their type.
+    if (data.isPostType) {
+      const { type, id } = data;
+      entity = state.source[type][id];
+    } else if (data.isTaxonomy) {
+      const { taxonomy, id } = data;
+      entity = state.source[taxonomy][id];
+    } else if (data.isAuthor) {
+      const { id } = data;
+      entity = state.source.author[id];
+    } else if (data.isPostTypeArchive) {
+      const { type } = data;
+      entity = state.source.type[type];
+    }
+
+    // It returns the entity found or `null` otherwise.
+    return entity;
   },
   data: {},
   category: {},